Update README.md
I wasn't quite sure how to edit the file, but I think it did it correctly. I haven't worked with arrays quite as much but I hope what I had type makes sense. trying to come up with a enhanced for loop for how it is going to go through a loop to check the availability.